# docs/swagger.yaml
basePath: /
definitions:
  models.AdCreate:
    properties:
      adset_id:
        type: string
      body_text:
        type: string
      headline:
        type: string
      image_url:
        type: string
      link_url:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  models.AdSetCreate:
    properties:
      billing_event:
        type: string
      campaign_id:
        type: string
      daily_budget:
        description: In paise/cents
        type: integer
      name:
        type: string
      optimization_goal:
        type: string
      status:
        type: string
      targeting:
        additionalProperties: true
        type: object
    type: object
  models.CampaignCreate:
    properties:
      name:
        type: string
      objective:
        type: string
      special_ad_categories:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  models.CampaignUpdate:
    properties:
      name:
        type: string
      objective:
        type: string
      status:
        type: string
    type: object
...
  models.MetaConnectRequest:
    properties:
      ad_account_id:
        description: e.g. "act_123456789"
        type: string
      catalog_id:
        type: string
      email:
        type: string
      page_id:
        type: string
      pixel_id:
        type: string
      short_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-Platform Ads Management API (Meta & Google) for Realify.
  title: Realify API
  version: "1.0"
paths:
...
  /meta/ads:
    post:
      consumes:
      - application/json
      description: Creates an ad with creative via the Meta API then persists it to
        the local database. Requires the user's Meta Page ID (resolved from X-User-Email
        or .env).
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Create ad
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/models.AdCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create a new ad
      tags:
      - meta
  /meta/ads/{id}/insights:
    get:
      description: Fetches live performance metrics for an ad directly from the Meta
        API (not local DB). Defaults to last 30 days if dates are omitted.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Ad ID
        in: path
        name: id
        required: true
        type: string
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get ad insights
      tags:
      - meta
  /meta/adsets:
    post:
      consumes:
      - application/json
      description: Creates an ad set within a campaign via the Meta API then persists
        it to the local database. Pass X-User-Email header to act as a specific user.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Create ad set
        in: body
        name: adset
        required: true
        schema:
          $ref: '#/definitions/models.AdSetCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create a new ad set
      tags:
      - meta
  /meta/adsets/{id}/ads:
    get:
      description: Fetch all ads belonging to a specific ad set via the Meta API.
        Pass X-User-Email header to act as a specific user.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Ad Set ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get ads in an ad set
      tags:
      - meta
  /meta/auth/token:
    post:
      consumes:
      - application/json
      description: Exchanges a short-lived Meta token for a long-lived one (60 days),
      parameters:
      - description: Connection payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.MetaConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Connect a Meta user (exchange token + save credentials)
      tags:
      - auth
  /meta/campaigns:
    get:
      description: Fetch all ad campaigns from the local database (results are synced
        from Meta). Pass X-User-Email header to query a specific user's campaigns.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all Meta campaigns
      tags:
      - meta
    post:
      consumes:
      - application/json
      description: Creates a campaign via the Meta API then persists it to the local
        database. Pass X-User-Email header to act as a specific user.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Create campaign
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/models.CampaignCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create a new Meta campaign
      tags:
      - meta
  /meta/campaigns/{id}:
    patch:
      consumes:
      - application/json
      description: Updates a campaign's status or name via the Meta API. Pass X-User-Email
        header to act as a specific user.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Update campaign
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/models.CampaignUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update a Meta campaign
      tags:
      - meta
  /meta/campaigns/{id}/insights:
    get:
      description: Fetches live performance metrics for a campaign directly from the
        Meta API (not local DB). Defaults to last 30 days if dates are omitted.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get campaign insights
      tags:
      - meta
  /meta/status:
    get:
      description: Check connection and list available pages/catalogs. Pass X-User-Email
        header to use a specific user's credentials; omits header to fall back to
        .env defaults.
      parameters:
      - description: User email (multi-user mode)
        in: header
        name: X-User-Email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Check Meta API connection
      tags:
      - meta
swagger: "2.0"